+++ packages/solid/src/methods/setError.ts
@@ -46,8 +46,9 @@ export function setError<
       // Update invalid state of form
       form.internal.setInvalid(
         !!error ||
-          [...form.internal.fields].some(
-            ([_, field]) => field.getActive() && field.getError()
+          [...form.internal.fields, ...form.internal.fieldArrays].some(
+            ([, fieldOrFieldArray]) =>
+              fieldOrFieldArray.getActive() && fieldOrFieldArray.getError()
           )
       );
     });